Hopefully bug fixed

+++ rvm/src/org/jikesrvm/runtime/SysCall.java
@@ -537,7 +537,7 @@ public abstract class SysCall {
 	   public abstract void print_logs();
 
 	   @SysCallTemplate
-	   public abstract void add_log_entry(double[] attrs,long cmdid,long timestamp,long freq);
+	   public abstract void add_log_entry(double[] attrs,int cmdid,long timestamp,int freq);
 	   @SysCallTemplate
 	   public abstract int add_method_entry(byte[] m, byte[] c);
